Move options to own file
